Moved auth functionality into services
